HNG-STAGE-2-TASK/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.3
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
